Adapt to the type information from Emacs 29.
Emacs 29 adds a new feature to bookmark-handler-functions: they can
carry symbol property named "bookmark-handler-type", which should be
recognized as the type of the bookmarks associated with such handlers.

This commit adds an automatic group that uses this type information
for classification, which works for a lot of cases out of the box, and
can be easily extended to handle custom cases as well.