Implement file-name-as-directory operation
From bug report #2
When calling

(file-relative-name buffer-file-name "/some/directory")
from a file opened with jarchive, file-relative-name is unable to build
a path relative to another directory. file-name-as-directory is called
repeatedly on the full `jar` or `zipfile` file URI, which always
includes the scheme. The path is never relative to a directory, which
keeps dwindling down the "/some/directory" -> "/some" -> "/" -> nil
until an error is thrown trying to check if the dwindled down path of
nil is a prefix of the full jar URI. This throws an exception. Having
`file-name-as-directory` operate on the jar path doesn't quite return a
correct answer (usually it returns `"."` in my testing), but the file in
the jar can't really be relative to anything on the filesystem, since
the jar acts like a remote file system with a separate file tree. This
at least doesn't throw an error.