Implement fallback to internal hashing function
When the checksum command cannot run because the executable is not available,
fallback to the internal hashing function `secure-hash'. When comparing
multiple locations, e.g. searching for files on local and remote hosts, the
executable might be usable on one host but not the other, so remember this per
host and try to be smart, because using the internal function works everywhere
but brings its own problems, like being generally slower (inserting into temp
buffer) and having the potential to exhaust memory when being too big.

Signed-off-by: Harald Judt <h.judt@gmx.at>