greader.el: «greader-continuous-mode' functionality added.
This minor-mode is for greader to automatically scroll in
some special types of buffers, like `info-mode' or external package
`nov-mode'.
For further details see the mode documentation.

In this version of the feature there is still a problem:
When using info-mode, the `greader-continuous-guess-function' function
correctly deduces the function to call for scrolling,
however it happens that in some specific pages in the manuals, the
function `Info-scroll-up' generates an "end-of-buffer" error. The cause
of this bizarre behavior is still unknown to me.