greader: version 0.9.4
General improvements to		`greader-dict--get-key-from-word':
Now it handles efficently the process of searching.
First, when there was not keys corresponding to a given word, the
algoritm scanned all the keys in the hash-table.
Now first the function checks if the word as is is present in the
hash-table, and in that case returns it.
If the word is not present as is, then we reduce the dictionary to
contain only the words classified as matches, and then searches for
the appropriate key.