greader: modified.
Improved management of sentences.
The addition of the `greader-dict.el' module made it necessary to modify
the way greader navigates between sentences.
Previously, the interface offered by 'sentence-forward' was used.
Now we use it, but by changing the `sentence-end' variable to one that
it's fine for greader, where very long sentences are not generated.