greader-dict.el: started implementation of filters.
filters.
filters allow users to define arbitrary regexps to be replaced
either with empty strings or by another string.
It is necessary to conceptually abstract filters from other types of
match because the filters allow the use of any character and in any case, being applied as they are,
the user can better exploit the expressive power of regexps.
so filters are a separate feature, which we can consider an
"advanced" use case of greader-dict.
Use the command `greader-dict-filter-add' which is not yet bound to
any key, as it is still an experimental feature.