Use multiple prompts for directories instead of completing-read-multiple
This commit gets rid of the usage of completing-read-multiple, which up until
the current release version of Emacs (29.2) does not work well with some
completing systems.

In addition to the possible buggy behaviour, using completing-read-multiple
might also not be a very obvious way for most users to specify multiple
directories, even if the possibility is mentioned in the README.

Thus, the new method is to ask the user for a directory, then for another
directory, and another one until the user enters nothing and just hits Enter.

Signed-off-by: Harald Judt <h.judt@gmx.at>