greader-speechd.el: Option `greader-speechd-handle-server' implemented.
This is, i hope, only a workaround while i discover the real problem
in the interaction between speech-dispatcher, spd-say and emacs.
The variable `greader-speechd-handle-server' enables greader to manage
the speech-dispatcher server. This implies that if there are an
instance of the server already running it will be killed and greader
will start a new one with the appropriate parameters to keep
speech-dispatcher running properly.
At this moment, if we don't enable this option, when we use
`greader-speechd' backend, the reading stops at the end of the first
sentence spoken.
I don't know why that happens, but maybe it has to do with timing
between spd-say, the client, and speech-dispatcher.
The reason for which i think so is that if we launch speech-dispatcher
with the flags -s or -d and we disable the timer with -t0, things work
properly and greader reads as usual.
But the idea of killing an instance of a process is very very bad, so
i hope to find another way to fix this.