* html/layout.css: (ol, ul) Enable bullets, numbers
This rule was causing ordered lists to be rendered without their
numbers.

I don't know why this rule was present; it may have been a stylistic
choice earlier in development, before ELPA was rendering packages'
readmes into the HTML, when it was likely that fewer types of elements
were present in ELPA's HTML, so the problem was not noticed.

Due to the way the rest of the page's layout is implemented, this
introduces a minor "ugliness": that, e.g. a numbered list's numbers
are aligned outside of the containing DIV.  This happens because all
of the DIVs on the page have their margin and padding set to
0, without which OLs and ULs would be naturally indented compared to
sibling paragraphs.  This will be handled in later commits which will
attempt to solve the margin/padding issues more "holistically."

(ol, ul): Delete list-style rule.