greader version 0.11.5
Customization option `greader-audiobook-include-track-name-in-audio'
added: This option allow us to have the title of the current track
spoken at the start.

Customization option `greader-audiobook-pause-at-end-of-track' added:
This option allow us to specify that we want an artificial pause at
end of  the current track maybe because owr reader is buggy, and jumps
if the track currently playing does'nt have a final pause.
You can also customize the variable `greader-audiobook-pause-string'
to change the string that is used to generate the pause.