files/emacs: project-dispatch: Reorganize commands in terms of locality
This organizes project-dispatch commands generally into sections in terms of
their options.  That is, the top-level ones which are always run at the root
directory are placed under "Project commands".  Going further down indicates
options like --from-{root,sub}-directory are relevant.

In the future, this structure should be made more apparent with descriptions
also reflecting the state (noted as TODO).

The "d" mapping seems fitting to bind to --from-{...}-directory, so the extra
dash has been removed to simplify keypresses.

Not all the "Find" commands can actually use subdirectories right now, but this
a technical issue, not a design one (i.e. needs to be programmed to work with
it).

Reorganization... again-again-again.