Retain modification mark when unmarking a marked book
* calibre-edit.el (calibre-edit--mark-modified): Rename calibre-edit-mark-modified.
(calibre-edit-apply):
(calibre-edit-add-tag):
(calibre-edit-remove-tag):

Adjust to renaming of calibre-edit--mark-modified.

(calibre-edit-modified-p): New function.
* calibre-library.el (calibre-library-mark-unmark): Preserve the 'M'
modification mark when unmarking a '*' mark.
* doc/calibre.texi: Remove warning about marking modified books.

If a book was modified and subsequently marked, unmarking it would
result in the modification mark being discarded.  This would result in
those modifications not being picked up by calibre-library-execute,
which would result in the modifications being lost.