Generalise calibre-edit-add/remove-tags
* calibre-edit.el (calibre-edit--metadata-adder):
  (calibre-edit--metadata-remover):
  (calibre-edit--metadata-modifier-pair):

  New macros.

  (calibre-edit-add-tags):
  (calibre-edit-remove-tags):

  Define using calibre-edit--metadata-modifier-pair.

The pair of functions to add to/remove from operations on any metadata
field which is a list will follow the same pattern as
calibre-edit-add-tags and calibre-edit-remove-tags with the only
difference being the metadata field being accessed.  By defining a
macro to generate such function pairs it becomes trivial to add the
ability to modify other metadata fields and ensures that all such
modification functions behave consistently.