emacs: project-dispatch: Fix current buffer always showing in buffer list.
The issue stems from DEFAULT-DIRECTORY being set on the current buffer, which
causes it to always be considered part of the project being visited, even when
it isn't.

This includes the environment-setting in the temporary buffer, which prevents
the current environment from being modified.

* files/emacs/impure/lisp/project-dispatch.el
(project-dispatch-switch-to-buffer): Swap WITH-TEMP-BUFFER and LET, so the
variables are only set in the temporary buffer.