emacs: project-dispatch: Support multiple compilation buffers.
Previously, we were constrained to one compilation running at a time since
attempting to run another command would conflict with the current buffer.
This adds a NAME to the compile-suffixes specification so that multiple
commands may run at the same time in their own buffers.

* files/emacs/impure/lisp/project-dispatch.el
(project-dispatch-compile-suffixes): Add NAME to the specification.
(project-dispatch-compile--setup-suffixes): Name compile buffers according to
the new specification.