Update disproject-compile-suffixes type to support additional properties
In order to support extending this in the future, we use plist at the end of
compile-suffix entries.  The description is now considered an optional property
that is specified via this mechanism.

These changes give the documentation for DISPROJECT-COMPILE-SUFFIXES a full pass
to expand on it and fix some errors.

For now, only include the compile command when description is not set.