Modify root directory query logic to be less "leaky"
Previously, this function was able to return a project root directory from both
the passed argument as well as `default-directory'.  This is somewhat leaky in
that the caller may be passing a directory specifically to avoid the function
using `default-directory', so we opt to only use one or the other.

More comments are also added to explain the logic behind the query ordering.