Transition to backing root directory only by Transient scope
Previously, there was not a One True Value for the root directory - it was
stored in both the Transient scope and as an infix option, and both relied on
each other in certain points of the code.

This simplifies the design so that the root directory value can be stored in a
single location while keeping other state intact (e.g. saving root directory
between prefixes; saving infix option state between switching projects).