Fix dir-locals prompt causing menu to break
Without this patch, if a user selects a project with potentially unsafe
directory-local variables (thus causing a prompt), the Transient menu doesn't
fully function anymore.  This seems to have to do with the prompt displaying
after Transient refreshes the menu, but the menu doesn't refresh again and gets
confused.

These changes check the Transient suffix related custom variables ahead of time
and put them in the scope so that the dir-locals prompt always happens earlier.

As a side effect, if the user does not wish to mark the variables as safe
permanently (e.g. selecting "y" instead of "!"), they should now only be
prompted at most once per menu.