Replace "directory" argument with "overrides" in scope setup
This allows for more flexibility in overriding values.

Corresponding invocations are modified to work with this new signature, and
prefix definitions are also changed to expect a project as the argument as well.
An exception is made for disproject-custom-dispatch, which has logic that is
actually not working as intended at the moment and needs additional changes out
of scope of this patch.

Docstring is also updated to include missing descriptions of scope values.