Fix selected project not being remembered in some cases
`project-remember' was only called for the active projects selection, but since
it was possible to pick an arbitrary directory in the prompt for
`disproject-switch-project', the project might not always be remembered.
Auto-selection on menu startup also bypasses this.

This moves the `project-remember' call to `disproject--setup-scope' which should
catch all relevant cases of "interacted with a particular project, so it should
be remembered".