Fix case of overwriting `project--list' and underlying file with `unset'
Normally, when performing actions that require interacting with the known
projects list, a function belonging to `project.el' that is used will call
`project--ensure-read-project-list', which initializes `project--list'.

When there are no active/open projects and no other `project.el' code has been
touched, it is possible for `project--list' to be uninitialized when invoking
`project-remember-projects-under' since the `seq-each' is completely skipped -
thus `project-remember-project' is never called, introducing an edge case where
the list is never loaded.

Since the call to `project--write-project-list' is unconditional, that means it
is possible for the /uninitialized value/ (i.e. `unset' symbol) to be written to
the backing file, wiping the known projects list.

The write function is now only called when the list of open projects is non-nil
to avoid this.