Rename `disproject-remember-projects-active'
...to `disproject-remember-projects-open'.

`disproject-remember-projects-active' is deprecated with this change.

The key-bind for remembering open active/open projects is also changed from
"r a" to "r o", with the former binding hidden and deprecated for removal.

This is part of a set of changes that change the references from "active
projects" to "open projects", which arguably makes more sense and follows
what Projectile calls them.