Replace alist-based scope with class-based system
This replaces the alist-based scope revolving around `disproject--setup-scope`
with a class-based approach, using `disproject-scope` as the top-level scope
type with `disproject-project` as a nested structure.

Some hacks are added to maintain support for preserving options across menus,
which will be resolved via other changes later on.

The signature for `disproject--scope` is changed to better fit the object in
scope now, since it assumed an alist before.

`disproject--state-*` functions are only changed internally to maintain
compatibility with code using them.