Remove `disproject--state-*` functions in favor of object interfaces
This layer of functions is arguably redundant, due to both the new
`disproject-{scope,project}` interfaces and also because they are not really
flexible.

The original purpose of these functions was for easily accessing transient
state - particularly infix arguments - with no-argument calls.  As such, this
usually involved returning nil on fail, which may not be desired in certain
cases.  However, with the increasing amount of state, it would be very easily to
balloon the amount of functions if we abided by the no-arguments rule for
transient scope queries; notably, most of these functions were used only once or
twice (or not at all).

These changes remove the layer in preference of using the object interfaces,
which encourage composing functions together to retrieve values while providing
more guarantees like ensuring non-nil return values.

In the case of prefix group declarations, the no-argument approach may be
preferable and could be added back for this specific purpose, but that is out of
scope for this patch.

Infix arguments may also still be satisfied by this approach.  An exception for
removal has been made for the `disproject--state-prefer-other-window?`, an infix
argument with numerous dependents that also needs refactoring.