Add compile command to main dispatch
This provides a compile command in `disproject-dispatch`, which serves to be an
easier-to-invoke command in the main dispatch and also to copy the `C-x p c`
key-bind that `project-compile` defaults to.

Since the "Alternative compile" suffix serves the same purpose, it is set to the
same identifier as this command to share process tracking.