Prevent cases where `disproject-process-buffer-name` can cause multiple prompts
Since the PROJECT-DIR argument was optional before, it was possible for two
separate scope objects to become out-of-sync and lead to undesired behavior from
ensuring a selected project more than once.

In the case of `disproject-process-buffer-name`, if the project was ensured via
`disproject-scope-selected-project-ensure`, but this function was called without
the PROJECT-DIR argument, it could result in prompting more than once for the
project directory since the scope objects may not necessarily be
shared (particularly when no transient scope is available).  By requiring the
project directory argument, we can avoid this.