User option `greader-reading-time-adjustment'.
This option allows to specify an adjustment in terms of percentage, to
fill time spent in pauses.
Some tests suggests that the overall time calculated by greader is the
10% less than the real time spent in reading the buffer.
It is necessary to make further testing to find a better way to handle
this issue.
Furthermore, mac-os back-end is somewhat slower than espeak, and the
overal pause time is around the 15%.
It may therefore be more functional to provide a per-back-end command
that returns the known percentage of
adjustment.