Remove summary as a metadata field
* calibre-book.el (calibre-book): Remove summary field.
* calibre-cli.el (calibre-cli--make-book): Don't retrieve the book's summary.
* calibre-core.el (calibre-library-columns): Remove summary as an option.
(calibre-book--print-info): Remove case for summary.
(calibre-book--file): Remove case for summary.
* calibre-db.el (calibre-db--make-book): Don't expect to receive summary.
(calibre-db--query): Don't retrieve the book's summary.

The summary field typically contains large amounts of text and does
not display well in the *Library* buffer.  This leads to a large
increase in memory consumption (it caused issues for `memory-report')
for very little, if any, gain.  For the upcoming feature for viewing
information on a single book we will be better served by fetching the
summary as needed.