blist.el: Refine the process of asking for tab names
Now the tab names are determined at the same time bookmarks are jumped
to, so that we do not risk opening the same bookmark twice, causing
some unwanted consequences.

Also the user experience of giving names to new tabs is refined as
well: the user can either choose manual names, the bookmark names, or
the automatic naming for the new tabs.  The user can even roll back a
tab if the user wants to undo some names.