Split up application of transient environment settings
This (similar to the previous commit) introduces functions and macros that split
up the functionality of `disproject-with-environment` under more clear semantics
of what should and shouldn't be considered applicable as advice.

A new function `disproject-with-env-apply` and its macro version
`disproject-with-env` are defined, which copy the behavior of
`disproject-with-environment` minus setting `default-directory`.

Transient advice is only applied when a suffix is invoked from a transient menu.
This makes sense for options like --prefer-other-window - which can only be
enabled through the menu - but does not for `default-directory` which was
/always/ assumed to be the project root when using
`disproject-with-environment`.  This has different semantics that make it unfit
to be applied in advice; i.e. `default-directory` may not actually be set to
the project root when a command is executed without a transient.

As such, we opt to exclude setting `default-directory` in
`disproject-with-env-apply`, and push it onto individual commands to explicitly
set it to the project root (perhaps with `disproject-with-project-root`).