Determine mode-line-process value on execution status not exec queue
* calibre-library.el (calibre-library-mode): Use
  calibre-exec--executing instead of calibre-exec--commands to
  determine whether to display ":Updating" in the mode line.

Using calibre-exec--commands leads to the ":Updating" string being
removed to early since the command queue becomes nil when the last
command starts execution.  calibre-exec--executing on the other hand
only becomes nil once the last command finishes execution.  It is also
probably the more semantically appropriate value to use since it is
explicitly the boolean value indicating whether a library update is in
progress.