Assume that the current transient suffix object is a valid type
The parent commit resolves issues with suffixes sometimes not having a nil
return value from `transient-suffix-object` even when in the context of an
interactive call to a suffix.  This means that an interactive call to a suffix
of the correct type will always have the expected slots that are referenced in
these affected suffixes, and permits assuming that the return value is valid.