Fix double-prompts to select project with shell command suffixes
We revert to arity 0 commands, with prompts occurring within command bodies
instead.  This avoids an issue due to state not being shared between the
interactive form and body, resulting in the selected project being discarded in
the initial interactive form (and thus another prompt when executing the body)
if called without a transient prefix.