Add `default-buffer-id` class slot to process suffixes
Determining that a default should be used requires the `buffer-id` slot value to
be nil, which made it infeasible to overload.  This adds a class slot to
`disproject-process-suffix`, which can be overridden per-subclass.

Buffer IDs for the main suffixes in prefix menus are still explicitly set so
that descriptions are not used, but are now adjusted to use buffer IDs from
corresponding suffix classes; e.g., `M-x disproject-compile` will now use the
same buffer as `M-x disproject-dispatch c`.