Improve advice management
* doc-dual-view.el (doc-dual-view--sync-pages): Only run
synchronization when mode is active.
(doc-dual-view--manage-advice): New function to centralize adding and
removing advice for all functions in doc-dual-view-modes.
(doc-dual-view--some-buffer-active-p): New function to check if any
buffer has the mode enabled.
(doc-dual-view-mode): Refactor to use the new helper functions and
only remove advice when no buffers are using the mode.