Add function to get book at point
* calibre-library.el (calibre-library-book-at-point): New function.
(calibre--get-active-books): Replace calls to tabulated-list-get-id
with calibre-library-book-at-point.
(calibre-library-mark-unmark): Replace calls to tabulated-list-get-id
with calibre-library-book-at-point.
(calibre-library-unmark-all): Replace calls to tabulated-list-get-id
with calibre-library-book-at-point.
(calibre-library-get-marked): Replace calls to tabulated-list-get-id
with calibre-library-book-at-point.
(calibre-library-execute): Replace calls to tabulated-list-get-id with
calibre-library-book-at-point.
(calibre-library-revert): Replace calls to tabulated-list-get-id with
calibre-library-book-at-point.
(calibre-library-open-book): Replace calls to tabulated-list-get-id
with calibre-library-book-at-point.
(calibre-library-open-book-other-window): Replace calls to
tabulated-list-get-id with calibre-library-book-at-point.
(calibre-library-open-book-external): Replace calls to
tabulated-list-get-id with calibre-library-book-at-point.

The use of tabulated-list-get-id to get the book at point is an
implementation detail that should be abstracted away.