fix: Allow directories to be passed to `vc-jj--filename-to-fileset`
There are various scenarios in which directories may be passed to
vc-jj; previously these cases were not handled properly.  As an
example, when calling `vc-next-action` on a vc-dir buffer with a
directory marked, the expected behavior is for all changed files in
that directory to be committed.  However, the actual behavior was that
none of these files would be committed.

The cause is in `vc-jj--filename-to-fileset`: it used the "cwd-file"
and "root-file" jj file patterns.  However, these file patterns do not
match files under directories.  Instead, to match files under
directories recursively, the "cwd" and "root" file patterns should be
used.

Fixes #62.