Add dispatch menu for finding common "special files"
Facilities for extending the menu more easily will be added later.

Remove comment about `disproject--custom-suffixes-prefix` being a hack, with the
addition of `disproject--selected-project-prefix`.  These'll remain private (for
the foreseeable future), but I think ensuring values in the scope like this is
actually the most reasonable place to do it.  It seems to be the
earliest (only?) exposed point where we can set a prefix's scope.  Doing it in a
prefix command's body means it's possible to unintentionally step around it if
setup occurs in another function; doing it any later also means we could run
into some kind of race condition where e.g. the scope is ensured for some
descriptions, but not for others.