fix: Make file paths relative to jj repo root
Normalize all file paths passed to `vc-jj--filename-to-fileset` to be
relative to the repository root.

Actually fixes #62, which was incorrectly thought to be fixed by
cca23877c1: users can now check in directories (all changed files
inside directories, recursively).  It is unclear why the previously
used "file:" is insufficient, but "root:" seems to remove the bug.