fix: Finding .git repo when inside .jj repo
When a .git repo is located in a subdirectory
within a .jj repo, previously project.el would
detect the current repo to be the further .jj
one instead of the closer .git one.

This commit fixes that bug while also simplifying
how vc-jj tells project.el whether we are within
a .jj repo.