Fix the file handler to also deal with insert-file-contents
The previous code was faulty. The idea is that we want to show preview
text only when we are visiting the file, but otherwise the functions
insert-file-contents and insert-file-contents-literally should work as
expected.

Thanks to Zhengyi Fu for reporting an error on the Emacs bug tracker
bug#78990: <https://lists.gnu.org/archive/html/bug-gnu-emacs/2025-07/msg00444.html>.