Add special handling to work with smtpmail
When sending mails, the auth-source queries from smtpmail doesn't
contain the :user field (meanwhile queries from Gnus seems to always
include :user).  When using predefined provider credentials, only the
:user field is different which is missing from the query, therefore we
should not just assume the matched result is the correct one.
Additionally, we need to check whether smtpmail-smtp-user, which was
set by smtpmail, is the same as :user to be sure.