Implement `vc-jj-root' using `vc-find-root' helper
This avoids starting a process. The only edge case I could think of is
the behaviour in the presence of backspaces. I verified that both `jj
root` and `(vc-find-root file ".jj")` return the same result. jj
creates a `.jj` directory on each workspace, vc-jj-root will return
the top-level directory of the workspace as the root, both in the
previous and the current version.

Copyright-paperwork-exempt: yes