Fix persist-default prop being set to current value
Previously, if a persist-defvar form was evaluated after the variable
was bound, the default value would be set to the whatever the current
value was instead of INITVALUE.  This sometimes led to a data loss
bug, since persist-save deletes when the current value is
persist-equal to the default value.  Resolves bug#75779.