Make denote-journal-calendar--date-to-identifier use the denote-journal-interval
Thanks to Abdelhak Bougouffa for noticing the error with the missing
argument. This was done in pull request 23, which technically was
fixing the problem, but I prefer this different style instead:
<https://github.com/protesilaos/denote-journal/pull/23>. The idea is
to have helper functions whose behaviour is easy to predict (granted,
we may not be doing it at all times).