Remove padding in default title format string
The %e format specifier inserts the day of the month padded with spaces, which results in a default format string like "Thursday  9 October 2025 14:14" (note the 2 spaces between Thursday and 9). Using %-e removes this extra padding, so that the behavior corresponds to that described in the doc string of the denote-journal-title-format variable.