greader-dict.el (greader-dict-add-entry): Prohibit punctuation in replacements
The validation for dictionary replacements was overly restrictive. It used
the "\\W" regexp class, which disallows any character that is not a
word constituent. This could prevent legitimate replacements.

The check is now more specific, using the "[[:punct:]]" character class
to only forbid punctuation marks. This aligns better with the intended
behavior. The user
error message has been updated accordingly.