Refactor: Simplify "greader-build-args" function
The "greader-build-args" function has been refactored to be more
concise and idiomatic.

The new implementation replaces the previous mix of "append", "cond",
and "dolist" with a more streamlined approach using "push", "when-let",
and "nreverse".

This change improves readability and maintainability by using
standard Elisp constructs for list manipulation. The removal of
"not-implemented" values is also handled more cleanly with "delete".