fix: Handle pushable bookmarks correctly in `vc-jj-bookmark-delete'
In Jujutsu, bookmarks marked with an asterisk (`*`) in the log
indicate local bookmarks that track a remote bookmark that points to a
different commit, i.e., bookmarks that can be pushed.

Previously, the names of these bookmarks in `vc-jj-bookmark-delete'
would erroneously include the asterisk, leading to attempts at
deleting bookmarks with incorrect names.  Do not include the asterisk
by using the .name() method on each bookmark name.