fix: `vc-jj-diff' edge case when REV1 is REV2
When calling `log-view-modify-change-comment', the diff associated
with that edit appears empty.

This is because Log View passes to `vc-jj-diff' the same revision for
its REV1 and REV2 arguments. As a result, "jj diff -f REV1 -t REV2" is
called.  There is a special case when REV1 is REV2: "jj diff" will
always return an empty commit.

Avoid this by calling "jj diff -r REV1" when REV1 equals REV2.