refactor: Internal process functions accept FILE-OR-LIST argument
`vc-jj--command-parseable`, `vc-jj--process-lines`, and
`vc-jj--command-dispatched` are the three internal functions we
currently use for invoking jj commands.  Unify calling conventions
such that each accepts a FILE-OR-LIST argument.  This centralizes
their logic and ensures that a fileset is passed to jj as opposed to
trusting the caller to convert files to fileset format.  (If the
caller needs to pass the raw filename instead of a fileset, then they
can do so using ARGS.)

Also clean up these functions' docstrings and make them more
informative.