Improve compatibility with certain oauth servers
* url-http-oauth.el (url-http-oauth--secret-auth-source): New
function.
(url-http-oauth--basic-auth): New function.
(url-http-oauth--get-access-token-grant): Make use of new
functions.
(url-http-oauth--save-bearer): Add optional argument for the
existing refresh token, and preserve it if a new refresh token is
not issued.
(url-http-oauth--refresh-access-token-grant): Include an
Authorization header with refresh requests.
(url-http-oauth--get-bearer): Retrieve the existing refresh token,
and pass it to url-http-oauth--save-bearer.
(url-oauth-auth): Encode the bearer token as utf-8 to fix an error
when making binary POST requests.